shaderapidx9/dx11: Use DirectX9Ex and DirectX 11 + com_ptr for no resource leaks
* Use DirectX9Ex by default as it is supported since WIndows 7
* Use FLIPEX swap effect by default as it is supported since Windows 7.
* Convert DirectX 10 renderer template to DirectX 11 one.
* Convert DirectX 10 shader proxy to DirectX 11 one.

+++ materialsystem/cmatnullrendercontext.cpp
@@ -346,7 +346,7 @@ public:
 	{
 	}
 
-	virtual int GetNumActiveDeformations() const
+	virtual intp GetNumActiveDeformations() const
 	{
 		return 0;
 	}
