+++ materialsystem/mat_stub.cpp
@@ -1734,7 +1734,7 @@ public:
 	{
 	}
 
-	virtual int GetNumActiveDeformations() const
+	virtual intp GetNumActiveDeformations() const
 	{
 		return 0;
 	}
